+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -608,7 +608,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KAB1TGN47491SVA9STS1GEM4"),
+  _id: z.string().default("01KAB6F73YN37XGZA2VZRTM6BF"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -651,7 +651,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KAB1TGN47491SVA9STS1GEM4"),
+  id: z.string().default("01KAB6F73YN37XGZA2VZRTM6BF"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,